feat: refactor membership API integration; consolidate API service imports and enhance configuration handling
