+++ .github/workflows/azure-static-web-apps-mango-bush-0b7a83b00.yml
@@ -4,10 +4,14 @@ on:
   push:
     branches:
       - deploy
+    paths:
+      - 'Client/**'
   pull_request:
     types: [opened, synchronize, reopened, closed]
     branches:
       - deploy
+    paths:
+      - 'Client/**'
 
 jobs:
   build_and_deploy_job:
